# Conflicts:
#	client-service/backend/client-service/src/main/java/com/cinema/clientservice/db/common/repositories/VersionOfferMovieMapRepository.java
#	client-service/backend/client-service/src/main/java/com/cinema/clientservice/db/instance/repositories/RepertoireRepository.java

+++ ticket-collector/frontend/TicketCollector/app/src/main/java/com/psi/ticket/collector/codeCheck/ui/CodeCheckScreen.kt
@@ -76,7 +76,7 @@ fun CodeCheckScreen(
                 value = text,
                 onValueChange = { code ->
                     text = code
-                    trigger(CodeCheckViewModel.Command.OnCodeChange(code.toLong()))
+                    trigger(CodeCheckViewModel.Command.OnCodeChange(code))
                 },
                 supportingText = {
                     if (viewState.isError) {
